Broad refactor of statsdash structure
Added new scheduling mechanism which allows us to define
reports/recipients/reporting frequency.  reports list in scheduler.py
acts as report configuration.

scheduler.py should be run hourly by cron. The scheduler checks if a
report can be run by querying the sqlite persistence layer.  It then
checks if the report's data is available and runs it/sends
appropriately.  It then persists the time of the run to sqlite so that
the next cron run won't pick this report up again.

reporting.py should broadly be considered as the controllers for
reports.  New reports can be defined here with templates, these should
query the Analytics object for their data.

analytics.py should be considered as the modelling layer - the Analytics
class should be the gateway to our data and should present methods for
generating reusable report data.

renderer.py is responsible for rendering django templates with provided
context.