Fixing weekly report generation schedule
Looks like the weekly schedule was running 7 days after the last run,
rather than respecting the 'weekday' frequency option.  This meant that
if sending was delayed - for any reason - the period of the report would
shift by a day.

This should now attempt to run on the weekday specified in the
report_schedule config.